server/requirements.txt
flask
flask_cors
langchain
openai
tiktoken
supabase
python-dotenv